requirements.txt
# Requirements - TP3 Vision Artificielle

# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0
torch>=1.11.0
torchvision>=0.12.0

# Data processing
pandas>=1.3.0
opencv-python>=4.5.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
jupyter>=1.0.0
ipython>=7.0.0

# Optional: Kaggle CLI
kaggle>=1.5.0

# Optional: GPU support
# cuda-runtime  # For NVIDIA GPU
